WebOfScience::Harvester - extract class refactor - WebOfScience::QueryAuthor
- consolidate author name-institution query in WebOfScience::QueryAuthor
  - only needs to return WOS-UIDs for an author
- delegate the WOS-queries to the WebOfScience.queries singleton
  - delegated methods for WebOfScience singleton can be private
